# api/openapi.yaml
openapi: "3.0.3"
info:
  title: SwayRider API
  version: "1.0"
  description: |
    Public HTTP API served by `swayrider-api` on port **8080**. It is the single
    entry point for mobile and web clients. Requests are proxied to backend
    microservices (authservice, regionservice, routerservice, searchservice,
    tilesservice) behind the scenes.

    ### Authentication

    Endpoints that require authentication accept the access token (JWT RS256,
    15-minute TTL) either as:
    - `Authorization: Bearer <token>` header, or
    - `access_token` HTTP-only cookie (base64-encoded, set by login/refresh).

    The refresh token (64-byte hex, 30-day TTL, single-use) is accepted via the
    `refresh_token` HTTP-only cookie or in the request body.

    ### SSE Protocol

    The `/api/v1/route`, `/api/v1/search`, and `/api/v1/search/reverse` endpoints
    deliver results asynchronously via **Server-Sent Events** (`text/event-stream`).

    Event types:

    | Event | Data |
    |---|---|
    | `queued` | `{"job_id":"…","queue_position":N}` |
    | `result` | Operation-specific result object |
    | `error` | `{"error":"…"}` |
servers:
  - url: https://api.swayrider.com
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
          description: JWT RS256 access token (15-minute TTL)
        refresh_token:
          type: string
          description: Hex-encoded 64-byte refresh token (30-day TTL, single-use)
      required: [access_token, refresh_token]
...
paths:
...
  /api/v1/auth/mfa/setup:
    post:
      summary: Start MFA enrollment
      description: |
        Generates a fresh TOTP secret for the authenticated user (stored encrypted,
        not yet enabled) and returns the base32 secret (shown once, for manual entry
        into an authenticator app), the otpauth URL, and a server-rendered QR PNG of
        the URL (base64).
      operationId: mfaSetup
      tags: [auth]
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Enrollment started
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                    description: Base32 TOTP secret, shown exactly once
                  otpauth_url:
                    type: string
                  qr_png_base64:
                    type: string
                    description: Base64-encoded PNG QR of the otpauth URL
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: MFA already enabled, or MFA disabled globally
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/mfa/enable:
    post:
      summary: Complete MFA enrollment
      description: |
        Verifies one TOTP code against the pending secret, enables MFA, and issues
        the fresh backup-code set (plaintext, shown once).
      operationId: mfaEnable
      tags: [auth]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: TOTP code from the authenticator app
              required: [code]
      responses:
        '200':
          description: MFA enabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  backup_codes:
                    type: array
                    items:
                      type: string
        '400':
          description: Invalid authentication code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: MFA already enabled, not set up, or disabled globally
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/mfa/disable:
    post:
      summary: Disable MFA
      description: |
        Disables MFA for the authenticated user (deletes the enrollment row and all
        backup codes). The account password is required and verified.
      operationId: mfaDisable
      tags: [auth]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  format: password
              required: [password]
      responses:
        '204':
          description: MFA disabled
        '401':
          description: Missing or invalid JWT, or wrong password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: MFA disabled globally
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/mfa/status:
    get:
      summary: Get MFA status
      description: Reports whether the authenticated user has MFA enabled.
      operationId: mfaStatus
      tags: [auth]
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: MFA status
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled:
                    type: boolean
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: MFA disabled globally
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/mfa/verify:
    post:
      summary: Complete the second factor
      description: |
        Exchanges the pending-login challenge token (from a login that returned
        `mfa_required`) plus a TOTP code or single-use backup code for the normal
        token pair. On success sets the same HTTP-only cookies as login.
      operationId: mfaVerify
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mfa_token:
                  type: string
                  description: Challenge token from the login response
                code:
                  type: string
                  description: TOTP code or single-use backup code
                remember_me:
                  type: boolean
                  default: false
              required: [mfa_token, code]
      responses:
        '200':
          description: Second factor accepted. Sets `access_token` and `refresh_token` cookies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Invalid, expired, or exhausted challenge; invalid code; locked MFA scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/mfa/backup-codes:
    post:
      summary: Regenerate backup codes
      description: |
        Replaces the authenticated user's backup-code set with a fresh one
        (invalidating the previous set) and returns the new plaintext codes once.
        The account password is required.
      operationId: mfaBackupCodes
      tags: [auth]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  format: password
              required: [password]
      responses:
        '200':
          description: New backup codes (shown once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  backup_codes:
                    type: array
                    items:
                      type: string
        '401':
          description: Missing or invalid JWT, or wrong password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: MFA not enabled, or disabled globally
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/auth/refresh:
    post:
      summary: Refresh access token
      description: |
        Exchange a refresh token for a new access + refresh token pair (token rotation).
        Accepts the refresh token from the `refresh_token` cookie (set by login) or
        from the request body.
      operationId: refreshToken
      tags: [auth]
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: Required if not using the cookie
                remember_me:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Token refreshed. Sets updated cookies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          description: Missing refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid, expired, or revoked refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/auth/request-password-reset:
    post:
      summary: Request a password reset email
      description: Always returns 204 to prevent email enumeration.
      operationId: requestPasswordReset
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                reset_url:
                  type: string
                  format: uri
              required: [email, reset_url]
      responses:
        '204':
          description: Request accepted (email sent if account exists)
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/reset-password:
    post:
      summary: Complete password reset
      operationId: resetPassword
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  format: uuid
                token:
                  type: string
                new_password:
                  type: string
                  format: password
              required: [user_id, token, new_password]
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired reset token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/verify-email:
    post:
      summary: Request a new verification email
      operationId: verifyEmail
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                verification_url:
                  type: string
                  format: uri
              required: [email, verification_url]
      responses:
        '204':
          description: Verification email sent
  /api/v1/auth/change-password:
    post:
      summary: Change the authenticated user's password
      operationId: changePassword
      tags: [auth]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                  format: password
                new_password:
                  type: string
                  format: password
              required: [old_password, new_password]
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Wrong current password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/check-password-strength:
    post:
      summary: Check password strength
      operationId: checkPasswordStrength
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
              required: [password]
      responses:
        '200':
          description: Strength result
          content:
            application/json:
              schema:
                type: object
                properties:
                  is_strong:
                    type: boolean
                  message:
                    type: string
...
  /api/v1/auth/admin/create-admin:
    post:
      summary: Create an admin user
      operationId: createAdmin
      tags: [admin]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required: [email, password]
      responses:
        '200':
          description: Admin created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                  message:
                    type: string
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...